moved charger slot creation from MTEBasicMachineWithRecipeBaseGui to parent class
